services/data-loader/requirements.txt
pandas==2.0.3
psycopg2-binary==2.9.9
